requirements.txt
dnspython
pymongo
python-dotenv
